# feat: improve performance of Resize