@disintegration 2018-10-16
Looking at the numbers, that's an impressive improvement in performance! Thank you!
I will review the changes this weekend.