@disintegration 2018-10-20
Merged. Thanks again for the contribution!